Move StatusNotifier infrastructure to platform
