+++ tests/applets/test_systemtray.py
@@ -4,17 +4,17 @@ from __future__ import annotations
 
 from gi.repository import GLib
 
-from docking.applets.systemtray.dbusmenu import parse_menu_node
 from docking.applets.systemtray.render import create_status_tray_icon
-from docking.applets.systemtray.state import (
+from docking.applets.systemtray.state import tooltip_text
+from docking.platform.status_notifier import (
     DEFAULT_ITEM_PATH,
     RegisteredItemAddress,
     StatusTrayState,
     parse_registered_item,
-    tooltip_text,
     tray_item_from_properties,
     unavailable_state,
 )
+from docking.platform.status_notifier.dbusmenu import parse_menu_node
 
 
 class TestRegisteredItemParsing:
